Move polls forms to separate file
